Add s.c.g.groupBy flag to group routes
Motivation: without groups, all the routes are persisted as part of the same set. If one route fails during /refresh, all the routes will be not be applied.

Having groups allows the developer to isolate by group of routes, so if one route is wrong it only impacts its group.

Also, if one route was persisted before the refresh of the route that fails, it will not be removed from routes even if the route belongs to the same group of the failing one.